requirements.txt
numpy==1.26.4
biopython==1.84
matplotlib==3.9.1
pytest==8.3.1
psutil==6.0.0
requests==2.32.3